numerics/init.h: max_n should have same type as taun

+++ numerics/init.h
@@ -36,7 +36,7 @@ struct initializations {
 
 	myfloat dtau;
 	myfloat tau_max;
-	unsigned long max_n;
+	int max_n;
 	myfloat max_wrongness;
 	myfloat min_x1;
 	myfloat max_x1;
